[DIAR-94] move: 파일 이름 변경
