+++ src/Components/Common/InputBox/InputBox.tsx
@@ -1,4 +1,4 @@
-import styles from './styles.module.scss';
+import styles from './InputBox.module.scss';
 
 export const InputBox = () => {
   return <div className={styles.fullContainer}>This is InputBox</div>;
