+++ src/Components/Common/TextArea/TextArea.tsx
@@ -1,4 +1,4 @@
-import styles from './styles.module.scss';
+import styles from './TextArea.module.scss';
 
 export const TextArea = () => {
   return <div className={styles.fullContainer}>This is TextArea</div>;
